requirements.txt
diffusers==0.25.0
Pillow==10.1.0
Pillow==10.1.0
torch==2.0.1+cu118
